+++ meta/main.yml
@@ -1,7 +1,7 @@
 ---
 galaxy_info:
-  author: "Ogonna Iwunze"
-  company: None
+  author: "O2 Priority"
+  company: "O2"
   license: BSD
   min_ansible_version: 2.4
   platforms:
